3_banque_dashboard/requirements.txt
# Étape 1 : Ingestion MongoDB
pymongo[srv]==4.6.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0

# Utilitaires généraux
numpy==1.26.3

# Étape 2 : Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Étape 3 : Extraction PDF
pdfplumber==0.11.0

# Étape 6 : Dashboard
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly>=6.1.1
statsmodels==0.14.0

# Étape 7 : Reporting (Conversion Notebook -> PDF via WebPDF)
nbconvert[webpdf]==7.16.4
nbformat==5.9.2
ipykernel==6.29.0
playwright==1.42.0
mistune==3.0.2
kaleido==1.2.0
